Restructure of ahm analytics tests
