--- tests/test_calculation_accuracy_metric.py
@@ -1,16 +1,0 @@
-import math
-
-import numpy as np
-
-from flownet.ahm._ahm_iteration_analytics import accuracy_metric
-
-
-def test_calculation_accuracy_metric() -> None:
-    assert math.isclose(
-        accuracy_metric(
-            np.array([[0.5, 0.5, 0.5], [0.5, 0.5, 0.5]]),
-            np.array([[0.4, 0.4, 0.4], [0.6, 0.6, 0.6]]),
-            "RMSE",
-        ),
-        0.1,
-    )
